Fix manifestwork and appliedmanifestwork unsync issue
merge the two controller as reconcilers for one controller
