+++ pkg/work/helper/helpers.go
@@ -139,6 +139,31 @@ func MergeStatusConditions(conditions []metav1.Condition, newConditions []metav1
 	return merged
 }
 
+// FindUntrackedResources returns applied resources which are no longer tracked by manifestwork
+// API version should be ignored when checking if a resource is no longer tracked by a manifestwork.
+// This is because we treat resources of same GroupResource but different version equivalent.
+// It also compares UID of the appliedResources to identify the untracked appliedResources because
+// 1. The UID should keep the same for resources with different versions.
+// 2. The UID in the newAppliedResources is always the latest updated one. The only possibility that UID
+// in appliedResources differs from what in newAppliedResources is that this resource is recreated.
+// Its UID in appliedResources is invalid hence recording it as untracked applied resource and delete it is safe.
+func FindUntrackedResources(appliedResources, newAppliedResources []workapiv1.AppliedManifestResourceMeta) []workapiv1.AppliedManifestResourceMeta {
+	var untracked []workapiv1.AppliedManifestResourceMeta
+
+	resourceIndex := map[workapiv1.ResourceIdentifier]struct{}{}
+	for _, resource := range newAppliedResources {
+		resourceIndex[resource.ResourceIdentifier] = struct{}{}
+	}
+
+	for _, resource := range appliedResources {
+		if _, ok := resourceIndex[resource.ResourceIdentifier]; !ok {
+			untracked = append(untracked, resource)
+		}
+	}
+
+	return untracked
+}
+
 // DeleteAppliedResources deletes all given applied resources and returns those pending for finalization
 // If the uid recorded in resources is different from what we get by client, ignore the deletion.
 func DeleteAppliedResources(
